requirements.txt
flask>=2.0.0
discord.py>=2.0.0
aiogram>=3.0.0b7
